fix weather lifecycle, get weather on load

+++ app/src/main/java/com/leandrososa/weatherapp/model/Coord.kt
@@ -3,7 +3,7 @@ package com.leandrososa.weatherapp.model
 import com.google.gson.annotations.SerializedName
 
 data class Coord(
-    @SerializedName("lat") val lat: Double,
-    @SerializedName("lon") val lon: Double
+    @SerializedName("lat") val lat: Float,
+    @SerializedName("lon") val lon: Float
 )
